[quant] Refactor to reduce the chances of cyclic imports
One of the reasons the cyclic imports happen when dealing with the quantization package is that there is a
tight coupling between the core torch.nn and torch.nn.quantized subpackages.
This diff starts the refactor to decouple the subpackages using an "interface":

Considering a hypothetical scenario, where the following sequence of events happen

1. `import torch` calls `from . import foo` (inside `torch/__init__.py`
1. `import foo` calls `import torch.nn.quantized`
1. `import torch.nn.quantized` imports all the modules
    1. `from modules import *` inside `torch/nn/quantized/__init__.py`
1. One of the modules in the `quantized` calls `import torch`

That scenario has an "import-time cyclic dependency`.
In practice that dependency might be a "run-time dependency", but due to the `from foo import *` mechanics,
the imports of the current package are not resolved until the imports of all subpackages in the stack
are not resolved.

Changes:

1. Move the `torch/nn/quantized/modules` to `torch/nn/quantized/_modules`.
    1. The contents of the `__init__.py` are empty to break any cyclic imports
1. Create an "interface" module `torch/nn/quantized/modules`
    1. The contents of the `__init__.py` define the user API of the subpackage
1. Move the `torch/nn/quantized/modules/utils.py` to `torch/nn/quantized/utils.py`
    1. This is done to avoid direct imports from `_modules`
1. Modify the `torch/ao/quantization/fx/_lower_to_native_backend.py` to use `_modules`

**Note:** If one needs to use the utilities in the subpackages, they must import from the `_modules`.
For example, if you are using `nnq.modules.conv._reverse_repeat_padding`, it must be changed to
`nnq._modules.conv._reverse_repeat_padding`.

[ghstack-poisoned]
