+++ torch/ao/nn/sparse/quantized/dynamic/linear.py
@@ -5,7 +5,7 @@ from torch.ao.nn.sparse.quantized.utils import LinearBlockSparsePattern
 
 import torch
 import torch.nn.intrinsic as nni
-from torch.nn.quantized.modules.utils import _quantize_weight, hide_packed_params_repr
+from torch.nn.quantized.utils import _quantize_weight, hide_packed_params_repr
 
 
 class Linear(torch.nn.Module):
