+++ torch/ao/nn/sparse/quantized/linear.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 import torch
-from torch.nn.quantized.modules.utils import _quantize_weight, hide_packed_params_repr
+from torch.nn.quantized.utils import _quantize_weight, hide_packed_params_repr
 
 # TODO (zaf): Inherit from `quantized.LinearPackedParams` (T83294430)
 class LinearPackedParams(torch.nn.Module):
