+++ torch/nn/quantized/_modules/conv.py
@@ -12,8 +12,9 @@ import torch.nn.intrinsic.qat as nniqat
 from torch._ops import ops
 from torch.nn.common_types import _size_1_t
 from torch.nn.modules.utils import _single, _pair, _triple
-from torch.nn.quantized.modules.utils import _quantize_weight, WeightedQuantizedModule
 from torch.nn.utils import fuse_conv_bn_weights
+
+from ..utils import _quantize_weight, WeightedQuantizedModule
 
 _SUPPORTED_PADDING = {
     'zeros',
