+++ torch/nn/quantized/_modules/embedding_ops.py
@@ -2,8 +2,9 @@ import torch
 import torch.nn as nn
 from torch import Tensor  # noqa: F401
 from torch._jit_internal import Optional, List  # noqa: F401
-from torch.nn.quantized.modules.utils import hide_packed_params_repr
-from torch.nn.quantized.modules.utils import _quantize_weight
+
+from ..utils import _quantize_weight
+from ..utils import hide_packed_params_repr
 
 class EmbeddingPackedParams(torch.nn.Module):
     _version = 1
