+++ torch/nn/quantized/_modules/linear.py
@@ -4,10 +4,11 @@ import torch
 import torch.nn as nn
 import torch.nn.intrinsic as nni
 import torch.nn.intrinsic.qat as nniqat
-from torch.nn.quantized.modules.utils import _quantize_weight, hide_packed_params_repr, WeightedQuantizedModule
 from torch.nn.utils.fusion import fuse_linear_bn_weights
 from torch.nn.utils.parametrize import type_before_parametrizations
 from typing import Optional
+
+from ..utils import _quantize_weight, hide_packed_params_repr, WeightedQuantizedModule
 
 class LinearPackedParams(torch.nn.Module):
     _version = 3
