+++ torch/nn/quantized/modules/__init__.py
@@ -1,17 +1,17 @@
 import torch
 from torch.nn.modules.pooling import MaxPool2d
 
-from .activation import ReLU6, Hardswish, ELU, LeakyReLU, Sigmoid, Softmax
-from .dropout import Dropout
-from .batchnorm import BatchNorm2d, BatchNorm3d
-from .normalization import LayerNorm, GroupNorm, InstanceNorm1d, \
+from .._modules.activation import ReLU6, Hardswish, ELU, LeakyReLU, Sigmoid, Softmax
+from .._modules.dropout import Dropout
+from .._modules.batchnorm import BatchNorm2d, BatchNorm3d
+from .._modules.normalization import LayerNorm, GroupNorm, InstanceNorm1d, \
     InstanceNorm2d, InstanceNorm3d
-from .conv import _ConvNd, Conv1d, Conv2d, Conv3d
-from .conv import ConvTranspose1d, ConvTranspose2d, ConvTranspose3d
-from .linear import Linear
-from .embedding_ops import Embedding, EmbeddingBag
+from .._modules.conv import _ConvNd, Conv1d, Conv2d, Conv3d
+from .._modules.conv import ConvTranspose1d, ConvTranspose2d, ConvTranspose3d
+from .._modules.linear import Linear
+from .._modules.embedding_ops import Embedding, EmbeddingBag
 
-from .functional_modules import FloatFunctional, FXFloatFunctional, QFunctional
+from .._modules.functional_modules import FloatFunctional, FXFloatFunctional, QFunctional
 
 
 class Quantize(torch.nn.Module):
